Observe characterData changes
This change is required because many sites update the DOM in such a way
that only this event is fired. I don't quite understand how (modifying a
text node rather than adding a new one?) but it happens just the same.

This did also require disconnecting and reconnecting appropriately to
avoid infinite recursion because the script would react to its own
changes.

Also, I factored out much of the logic for dealing with updated nodes
into its own function.

This patch solves the issues with Reddit.